+++ addons/binding/org.openhab.binding.pioneeravr/src/main/java/org/openhab/binding/pioneeravr/internal/handler/AbstractAvrHandler.java
@@ -26,20 +26,20 @@ import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandler;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.UnDefType;
-import org.openhab.binding.pioneeravr.PioneerAvrBindingConstants;
+import org.openhab.binding.pioneeravr.internal.PioneerAvrBindingConstants;
 import org.openhab.binding.pioneeravr.internal.protocol.RequestResponseFactory;
-import org.openhab.binding.pioneeravr.protocol.AvrConnection;
-import org.openhab.binding.pioneeravr.protocol.AvrConnectionException;
-import org.openhab.binding.pioneeravr.protocol.AvrResponse;
-import org.openhab.binding.pioneeravr.protocol.CommandTypeNotSupportedException;
-import org.openhab.binding.pioneeravr.protocol.event.AvrDisconnectionEvent;
-import org.openhab.binding.pioneeravr.protocol.event.AvrDisconnectionListener;
-import org.openhab.binding.pioneeravr.protocol.event.AvrStatusUpdateEvent;
-import org.openhab.binding.pioneeravr.protocol.event.AvrUpdateListener;
-import org.openhab.binding.pioneeravr.protocol.states.MuteStateValues;
-import org.openhab.binding.pioneeravr.protocol.states.PowerStateValues;
-import org.openhab.binding.pioneeravr.protocol.utils.DisplayInformationConverter;
-import org.openhab.binding.pioneeravr.protocol.utils.VolumeConverter;
+import org.openhab.binding.pioneeravr.internal.protocol.avr.AvrConnection;
+import org.openhab.binding.pioneeravr.internal.protocol.avr.AvrConnectionException;
+import org.openhab.binding.pioneeravr.internal.protocol.avr.AvrResponse;
+import org.openhab.binding.pioneeravr.internal.protocol.avr.CommandTypeNotSupportedException;
+import org.openhab.binding.pioneeravr.internal.protocol.event.AvrDisconnectionEvent;
+import org.openhab.binding.pioneeravr.internal.protocol.event.AvrDisconnectionListener;
+import org.openhab.binding.pioneeravr.internal.protocol.event.AvrStatusUpdateEvent;
+import org.openhab.binding.pioneeravr.internal.protocol.event.AvrUpdateListener;
+import org.openhab.binding.pioneeravr.internal.protocol.states.MuteStateValues;
+import org.openhab.binding.pioneeravr.internal.protocol.states.PowerStateValues;
+import org.openhab.binding.pioneeravr.internal.protocol.utils.DisplayInformationConverter;
+import org.openhab.binding.pioneeravr.internal.protocol.utils.VolumeConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
