+++ addons/binding/org.openhab.binding.pioneeravr/src/main/java/org/openhab/binding/pioneeravr/internal/protocol/utils/DisplayInformationConverter.java
@@ -10,9 +10,9 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.pioneeravr.protocol.utils;
+package org.openhab.binding.pioneeravr.internal.protocol.utils;
 
-import org.openhab.binding.pioneeravr.protocol.AvrConnectionException;
+import org.openhab.binding.pioneeravr.internal.protocol.avr.AvrConnectionException;
 
 /**
  *
